The DOM module states its own platform shape
A module whose one reason was conformance dissolves: the ledger is the
last section of dom.ts, the classes' own closing statement about how
far from lib.dom they stand. Installer-owned bins keep leaving for the
files that install them, the way the engine's geometry already did;
what remains here shrinks toward the gaps and the nevers.
